Make bootstrap scripts idempotent with version checking
Add version detection to Go, Rust, Neovim, and Alacritty installation scripts to skip reinstallation when target versions are already installed. Remove unused local/bin directory creation from environment setup.

+++ bootstrap/20-install-neovim.sh
@@ -11,24 +11,35 @@ npm install --global tree-sitter
 # For telescope
 sudo apt install --yes ripgrep
 
-sudo apt --yes install \
-  ninja-build \
-  gettext \
-  libtool \
-  libtool-bin \
-  autoconf \
-  automake \
-  cmake \
-  g++ \
-  pkg-config \
-  unzip \
-  doxygen
-
 NEOVIM_VERSION=$(get_latest_version neovim/neovim)
 
-rm -rf ${HOME}/src/neovim
-git clone https://github.com/neovim/neovim.git ${HOME}/src/neovim
-cd ${HOME}/src/neovim
-git checkout v${NEOVIM_VERSION}
-make CMAKE_BUILD_TYPE=Release
-sudo make install
+current_nvim_version=""
+if command -v nvim &> /dev/null; then
+    current_nvim_version=$(nvim --version | head -n1 | grep -oE '[0-9]+\.[0-9]+\.[0-9]+' || echo "")
+fi
+
+if [ "$current_nvim_version" = "$NEOVIM_VERSION" ]; then
+    echo "Neovim v$NEOVIM_VERSION is already installed"
+else
+    echo "Installing Neovim v$NEOVIM_VERSION (current: ${current_nvim_version:-none})"
+    
+    sudo apt --yes install \
+      ninja-build \
+      gettext \
+      libtool \
+      libtool-bin \
+      autoconf \
+      automake \
+      cmake \
+      g++ \
+      pkg-config \
+      unzip \
+      doxygen
+
+    rm -rf ${HOME}/src/neovim
+    git clone https://github.com/neovim/neovim.git ${HOME}/src/neovim
+    cd ${HOME}/src/neovim
+    git checkout v${NEOVIM_VERSION}
+    make CMAKE_BUILD_TYPE=Release
+    sudo make install
+fi
